+++ scripts/__tests__/__testfixtures__/Button.input.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { TouchableOpacity, Text } from 'react-native';
+
+interface ButtonProps {
+  text: string;
+  onPress: () => void;
+}
+
+export const Button = ({ text, onPress }: ButtonProps) => (
+  <TouchableOpacity onPress={onPress}>
+    <Text>{text}</Text>
+  </TouchableOpacity>
+);